Serialize element info.
